created database file to handle local data
